\name{fullrks}
\alias{fullrks}
\title{Internal functions of TWIX.}
\description{
Internal functions of TWIX.
}
\usage{
fullrks(m)
}
\arguments{
\item{m}{a numeric or character vector}
}
\examples{
fullrks(c(3,1,2,4,5))
a<-as.factor(c("yes","yes","no","no"))
fullrks(a)
}
\keyword{tree}
