\name{nlogn}
\alias{nlogn}
\title{Internal functions of TWIX.}
\description{
This function compute \code{0*log(0)}.
}
\usage{
nlogn(x)
}
\arguments{                                                                           
\item{x}{ a numeric value }
}
\examples{
log(0)
nlogn(0)
}
\keyword{tree}
