\name{olives}
\alias{olives}
\title{Classification of olive oils from their fatty acid composition.}
\description{
  The \code{olives} data frame has 572 rows and 10 columns.
}
\usage{
olives
}
\format{
This data frame contains the following columns:\cr
\code{Area} a factor with 3 levels.\cr
\code{Region} a factor with 9 levels.\cr
and 8 variables of fatty acid measurements.
}
\source{
   	Forina, M. and Armanino, C. and Lanteri, S. and Tiscornia, E. (1983)\cr
   	\emph{Food Research and Data Analysis},\cr 
	Applied Science Publishers, CA 1983.
}
\keyword{datasets}
