\name{pred.value}
\alias{pred.value}
\title{Prediction for one case}
\description{
 Internal prediction functions.
}
\usage{
pred.value(x, tree)
}
\arguments{                                                                           
\item{x}{ the case from data.frame}  
\item{tree}{ an object of class \emph{single.tree}}                                                                     
}                                                                                     
\seealso{
  \code{\link{predict.TWIX}}  
}                        
\keyword{tree}
