\name{print.id.tree}
\alias{print.id.tree}
\title{
Internal function of TWIX.
}
\description{
    Print names from \code{id.tree} object.}
\usage{
\method{print}{id.tree}(x, sq = 1:5, ...)
}
\arguments{
  \item{x}{an object of class \code{id.tree}.}
  \item{sq}{a numeric vector.}
  \item{\dots}{further arguments passed to or from other methods.}
}
\seealso{
  \code{\link{TWIX}}}
\keyword{tree}
