\name{summary.TWIX}
\alias{summary.TWIX}
\alias{summary.bootTWIX}
\title{Summarising TWIX}
\description{
  \code{summary} method for objects returned by \code{\link{TWIX}} or
\code{bootTWIX}.}
\usage{
\method{summary}{TWIX}(object, \dots)
\method{summary}{bootTWIX}(object, \dots)
}
\arguments{
  \item{object}{object of class \code{TWIX} or \code{bootTWIX}.}
  \item{\dots}{further arguments to be passed to or from methods.}
}
\seealso{                                                  
  \code{\link{TWIX}}
}
\keyword{tree}
