\name{print.TWIX}
\alias{print.TWIX}
\alias{print.bootTWIX}
\title{Print method for TWIX or BootTWIXtree object}
\description{
Print object of class \code{TWIX} or \code{bootTWIX}.
}
\usage{
\method{print}{TWIX}(x,...)
\method{print}{bootTWIX}(x,...)
}
\arguments{
  \item{x}{object of class \code{TWIX} or \code{bootTWIX}.}
  \item{\dots}{additional arguments.}
}
\details{
An object of class \code{TWIX} or \code{bootTWIX} is printed.
}
\seealso{                                              
  \code{\link{TWIX}}, \code{\link{print.single.tree}}   
}                                                      
\keyword{tree}
